% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GD2Viz-pkg.R
\docType{package}
\name{GD2Viz-pkg}
\alias{GD2Viz-package}
\alias{GD2Viz-pkg}
\title{GD2Viz}
\description{
TODO DESCRIPTION
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/yourusername/GD2Viz}
  \item Report bugs at \url{https://github.com/yourusername/GD2Viz/issues}
}

}
\author{
Arsenij Ustjanzew \email{arsenij.ustjanzew@uni-mainz.de}

Maintainer: Arsenij Ustjanzew \email{arsenij.ustjanzew@uni-mainz.de}
}
