% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{GD2Viz}
\alias{GD2Viz}
\title{GD2Viz main function}
\usage{
GD2Viz()
}
\arguments{
\item{TODO}{a \code{data.frame}, The input data used for GeDi. This should be
a \code{data.frame} of at least two columns.}
}
\value{
A Shiny app object is returned
}
\description{
GD2Viz main function
}
\examples{
if (interactive()) {
  GD2Viz()
}

}
