% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getEdgeWeightNStepsAway}
\alias{getEdgeWeightNStepsAway}
\title{Get Edge Weight N Steps Away}
\usage{
getEdgeWeightNStepsAway(g, start_col, start_row)
}
\arguments{
\item{g}{A numeric matrix representing the adjacency matrix of the graph. The matrix should contain edge weights as its values.}

\item{start_col}{An integer specifying the column index of the starting node.}

\item{start_row}{An integer specifying the row index of the starting node.}
}
\value{
If the edge is not reachable due to no incoming edges, the function returns NULL. If the edge is reachable, the function returns the weight of the edge that is a certain number of steps away.
}
\description{
The \code{getEdgeWeightNStepsAway} function finds the weight of an edge that is a certain number of steps away from a specified node in a graph. This function is used within the context of the \code{computeTransitionProbability} function to calculate transition probabilities along paths.
}
\note{
The function assumes that the input graph is represented as an adjacency matrix with non-negative edge weights.
}
\examples{
# Create a sample adjacency matrix
g <- matrix(c(0, 1, 2, 0, 0, 0, 1, 0, 0), nrow = 3, byrow = TRUE)

# Get the weight of an edge n steps away
weight <- getEdgeWeightNStepsAway(g, 2, 3)
}
