% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normalizeTestDataset}
\alias{normalizeTestDataset}
\title{Estimate Size Factors for New Dataset Based on Training Geometric Means}
\usage{
normalizeTestDataset(counts = NULL, metadata = NULL, dds = NULL, geom)
}
\arguments{
\item{counts}{A data frame or matrix containing raw unnormalized RNA-Seq gene counts. Columns are samples, rows are genes.}

\item{metadata}{A data frame containing sample metadata. Required if counts is provided.}

\item{dds}{A DESeqDataSet object. If provided, counts and metadata are ignored.}

\item{geom}{A named numeric vector containing the geometric means of the training dataset.}
}
\value{
A DESeqDataSet object with estimated size factors.
}
\description{
This function estimates size factors for a new dataset based on the geometric means from a training dataset.
It can take either a counts table with metadata or a DESeqDataSet object.
}
\examples{
\dontrun{
geom <- readRDS("./SVM_GD2_dashboard/data/geom_train_data.Rds")

counts <- read.table("SVM_GD2_dashboard/data/datasets/rus_cell_lines_counts.tsv", sep = "\t", header = TRUE)
metadata <- read.table("SVM_GD2_dashboard/data/datasets/rus_cell_lines_metadata.tsv", sep = "\t", header = TRUE)

dds <- normalizeTestDataset(counts = counts, metadata = metadata, geom = geom)
}
}
