% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{range01}
\alias{range01}
\title{Rescale Vector to Range 0-1}
\usage{
range01(x, ...)
}
\arguments{
\item{x}{A numeric vector to be rescaled.}

\item{...}{Additional arguments to be passed to \code{min} and \code{max} functions.}
}
\value{
A numeric vector with values rescaled to the range [0, 1].
}
\description{
The \code{range01} function rescales a vector to the range [0, 1].
}
\examples{
# Rescale a vector to the range [0, 1]
x <- c(1, 2, 3, 4, 5)
range01(x)

}
