% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{resample}
\alias{resample}
\title{Resample Function}
\usage{
resample(x, ...)
}
\arguments{
\item{x}{A vector from which elements will be sampled.}

\item{...}{Additional arguments to be passed to \code{sample.int}.}
}
\value{
A vector containing a random sample of elements from \code{x}, with replacement.
}
\description{
The \code{resample} function randomly selects elements from a vector with replacement.
}
\details{
This function is a wrapper around the \code{sample.int} function in R, providing a more concise way to resample elements from a vector.
}
\examples{
# Resample elements from a vector
x <- c(1, 2, 3, 4, 5)
resample(x, 3)

}
\seealso{
\code{\link{sample.int}}
}
