% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{trainGD2model}
\alias{trainGD2model}
\title{Train GD2 Model}
\usage{
trainGD2model(
  train_data,
  adjust_ras = c("ras", "ras_prob", "ras_prob_rec"),
  adjust_input = c("raw", "ranged", "scaled"),
  center = TRUE,
  type_column = "X_study"
)
}
\arguments{
\item{train_data}{A SummarizedExperiment object containing training data. It should include:
\itemize{
\item assays: "ras", "ras_prob", "ras_prob_path", "ras_prob_rec".
\item colData: Column data information.
\item metadata(train_data)$geom: A numeric vector of geometric mean data for genes.
}}

\item{adjust_ras}{Which assay to use. One of "ras", "ras_prob", "ras_prob_path", or "ras_prob_rec".}

\item{adjust_input}{How to adjust the input matrix: "raw", "ranged", or "scaled".}

\item{center}{Logical; whether to mean-center data if using "scaled" adjustment.}

\item{type_column}{Column name in colData indicating the sample type.}
}
\value{
An SVM model trained using kernlab::ksvm.
}
\description{
The \code{trainGD2model} function trains a GD2 model using a specified reaction activity score (RAS) matrix from training data
stored as a SummarizedExperiment object.
}
\examples{
\dontrun{
# Train GD2 model with raw input
model <- trainGD2model(train_data, adjust_ras = "ras", adjust_input = "raw")
}
}
