% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GD2Viz-pkg.R
\docType{package}
\name{GD2Viz-pkg}
\alias{GD2Viz-package}
\alias{GD2Viz-pkg}
\title{GD2Viz}
\description{
Welcome to \strong{GD2Viz}, a powerful and user-friendly visualization tool designed to help researchers and clinicians explore, analyze, and interpret GD2 Scores across various RNA-Seq datasets. Utilizing the advanced methodology outlined by Ustjanzew et al. (2024), \strong{GD2Viz} offers an in-depth examination of precomputed GD2 Scores from publicly available RNA-Seq datasets such as TCGA, GTEx, and TARGET. Users also have the flexibility to upload and analyze their own datasets within the context of GD2 Scores. This R Shiny application version of \strong{GD2Viz} provides a range of interactive visualizations and preloaded datasets. Additionally, the \strong{GD2Viz} R package includes several functions for computing Reaction Activity Scores of glycosphingolipid metabolism and predicting GD2 Scores directly within the R environment.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/arsenij-ust/GD2Viz}
  \item \url{https://arsenij-ust.github.io/GD2Viz/}
  \item Report bugs at \url{https://github.com/arsenij-ust/GD2Viz/issues}
}

}
\author{
Arsenij Ustjanzew \email{arsenij.ustjanzew@uni-mainz.de}

Maintainer: Arsenij Ustjanzew \email{arsenij.ustjanzew@uni-mainz.de}
}
