% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{trainGD2model}
\alias{trainGD2model}
\title{Train GD2 Model}
\usage{
trainGD2model(
  train_data,
  adjust_ras = c("ras", "ras_prob", "ras_prob_path", "ras_prob_rec"),
  adjust_input = c("raw", "ranged", "scaled"),
  center = TRUE,
  type_column = "X_study"
)
}
\arguments{
\item{train_data}{A list containing the training data. It should include:
\itemize{
\item \code{geom} - Geometric information.
\item \code{coldata} - Column data information.
\item \code{ras} - A matrix of reaction activity scores for key reactions.
\item \code{ras_prob} - A matrix of reaction activity scores with probabilities.
\item \code{ras_prob_path} - A matrix of pathway-specific reaction activity scores with probabilities.
\item \code{ras_prob_rec} - A matrix of receptor-specific reaction activity scores with probabilities.
}}

\item{adjust_ras}{A character string indicating which RAS matrix to use from \code{train_data}. Options are "ras", "ras_prob", "ras_prob_path", and "ras_prob_rec". Default is "ras".}

\item{adjust_input}{A character string indicating how to adjust the input data.
Options are "raw" (no adjustment), "ranged" (rescale to [0, 1]), and "scaled" (standardize with mean 0 and standard deviation 1). Default is "raw".}

\item{center}{TODOTODO}

\item{type_column}{TODOTODO}
}
\value{
An SVM model trained on the input and output sums derived from the RAS data.
}
\description{
The \code{trainGD2model} function trains a GD2 model using a specified reaction activity score (RAS) matrix from training data.
It allows for various adjustments to the input data before training and can rescale the output.
}
\examples{
\dontrun{
# Train GD2 model with raw input
model <- trainGD2model(train_data, adjust_ras = "ras", adjust_input = "raw")
}
}
